Preparation for future statistics collection
Such as:
- Total donations, average donations, donations by rank
- How many servers there are on each rank
- Average playing time on each rank
- Etc.
